2024/10/06:  Adds plugin for Emblem Trader
