+++ src/main/java/io/xeros/model/Shops.java
@@ -13,4 +13,5 @@ public class Shops {
     public static final Integer NURMOF_SHOP = 90;
     public static final Integer SIGMUND_THE_MERCHANT_SHOP = 26;
     public static final Integer HUNTER_SHOP = 23;
+    public static final Integer EMBLEM_TRADER_SHOP = 80;
 }
